Introduce ModelFactory to be able to cache file contents of translation unit
It seems the file contents is not available in libclang entites until
we have actually traversed the AST a bit. Thus we extract the source
file contents for the translation unit when accessing the first
entity.

The file contents is cached in the new class `ModelFactory` used
instead of having methods directly in `ClassToMock`.